MNT: fix cyclic import
