+++ rocketpy/mathutils/function.py
@@ -17,6 +17,7 @@ from pathlib import Path
 
 import matplotlib.pyplot as plt
 import numpy as np
+from numpy import trapezoid
 from scipy import integrate, linalg, optimize
 from scipy.interpolate import (
     LinearNDInterpolator,
@@ -27,14 +28,6 @@ from scipy.interpolate import (
 
 from rocketpy.plots.plot_helpers import show_or_save_plot
 from rocketpy.tools import deprecated, from_hex_decode, to_hex_encode
-
-# Numpy 1.x compatibility,
-# TODO: remove these lines when all dependencies support numpy>=2.0.0
-if np.lib.NumpyVersion(np.__version__) >= "2.0.0b1":
-    # pylint: disable=no-name-in-module
-    from numpy import trapezoid  # pragma: no cover
-else:
-    from numpy import trapz as trapezoid  # pragma: no cover
 
 NUMERICAL_TYPES = (float, int, complex, np.integer, np.floating)
 INTERPOLATION_TYPES = {
